+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Version 0.6.6 - 2026-04-22
 
-- Add predictions_format argument to `get_prediction`
-- Get predictions with correct output format in `create_prediction`
+- Get predictions with correct output format in `create_prediction` and `get_prediction`
 
 ## Version 0.6.5 - 2026-04-21